Remove useless components and props
