+++ client/sites-dashboard/components/sites-grid-item/README.md
@@ -1,11 +1,12 @@
 # SitesGridItem
 
-Renders a SitesGridItem component.
+Renders a SitesGridItem component with site selection option.
 
 ## How to use
 
 ```jsx
 import SitesGridItem from 'calypso/sites-dashboard/components/sites-grid-item';
+import type { SiteExcerptData } from '@automattic/sites';
 
 function render() {
 	const site = {};
@@ -14,7 +15,8 @@ function render() {
 			<SitesGridItem
 				site={ site }
 				key={ site.ID }
-			></SitesGridItem>
+				onSiteSelectBtnClick={ ( s: SiteExcerptData ) => {} }
+			/>
 		</div>
 	);
 }
@@ -24,8 +26,4 @@ function render() {
 
 - `site`: a site data e.g. SiteExcerptData object.
 - `key`: unique key eg. Site ID.
-- `showLaunchNag`: boolean, optional, default: true.
-- `showBadgeSection`: boolean, optional, default: true.
-- `showThumbnailLink`: boolean, optional, default: true.
-- `showSiteRenewLink`: boolean, optional, default: true.
-- `onSiteSelectBtnClick`: function, optional, default: undefined.
+- `onSiteSelectBtnClick`: function.
