+++ client/sites-dashboard/utils.ts
@@ -106,10 +106,6 @@ export const generateSiteInterfaceLink = (
 	return targetLink;
 };
 
-export const getSiteWpAdminUrl = ( site: SiteExcerptNetworkData ) => {
-	return site?.options?.admin_url ?? '';
-};
-
 export const SMALL_MEDIA_QUERY = 'screen and ( max-width: 600px )';
 
 export const MEDIA_QUERIES = {
